Add a clock to SelectMusic, date on Evaluation and EvaluationSummary
As long as you are in home mode (with or without Event Mode enabled), a clock will be shown on the bottom bar of the theme on select screens. On other screens with screenshot capability, this will show a date timestamp in the Wendy font instead (looks better in Wendy on those screens). Turning on Event Mode while still in Pay/Free mode will not show the clock. The clock is intended for home users only because of a lack of space in paid modes (and not to be confusing with the possibility of a menu timer). In the future I may make it configurable to display either 24-hour or 12 hour time, but as it is now, this displays 24-hour time.

This also slightly adjusts the positioning on ScreenEvaluation and changes the clock format to "Month Date, Year"