Add quick restart to ScreenGameplay (CTRL+R)
To use this, enable Event Mode. When on ScreenGameplay, press [Left CTRL + R]. The song will completely restart. This will work independently of "Action On Missed Target", which is an advanced player modifier.

This is an adaptation of TaroNuke's restart module from ECFA2021. His initial implementation has a bug whereby if you're in anything other than event mode, trigger the restart, then back out with the ESC key, the MusicWheel is empty. Rather than fix his implementation for use outside of Event Mode, I'm going to add a check to make sure that event mode is enabled. Let's be honest, as much as I abused free credit codes, being able to endlessly restart a song in a paid environment is a bit much😉.

Co-Authored-By: TaroNuke <633391+TaroNuke@users.noreply.github.com>